Fixing warnings in MotionControllerModel.cpp

+++ UnityEditorMotionControllerModelDLL/MotionControllerModel/MotionControllerModel.cpp
@@ -64,8 +64,8 @@ EXTERN_C __declspec(dllexport) BOOL TryGetMotionControllerModel(_In_ UINT32 cont
                 }
 
                 // Create a buffer from the stream, to read the contents into a byte array for passback to the app.
-                IBuffer^ buffer = ref new Buffer(stream->Size);
-                auto readOperation = create_task(stream->ReadAsync(buffer, stream->Size, InputStreamOptions::None));
+                IBuffer^ buffer = ref new Buffer(static_cast<unsigned int>(stream->Size));
+                auto readOperation = create_task(stream->ReadAsync(buffer, static_cast<unsigned int>(stream->Size), InputStreamOptions::None));
 
                 status = readOperation.wait();
